Make times `unsigned long` to avoid overlow
on 16 bit architectures. Closes #10

+++ MHZ.cpp
@@ -8,11 +8,11 @@
 const int MHZ14A = 14;
 const int MHZ19B = 19;
 
-const int MHZ14A_PREHEATING_TIME = 3 * 60 * 1000;
-const int MHZ19B_PREHEATING_TIME = 3 * 60 * 1000;
+const unsigned long MHZ14A_PREHEATING_TIME = 3 * 60 * 1000;
+const unsigned long MHZ19B_PREHEATING_TIME = 3 * 60 * 1000;
 
-const int MHZ14A_RESPONSE_TIME = 60 * 1000;
-const int MHZ19B_RESPONSE_TIME = 120 * 1000;
+const unsigned long MHZ14A_RESPONSE_TIME = 60 * 1000;
+const unsigned long MHZ19B_RESPONSE_TIME = 120 * 1000;
 
 const int STATUS_NO_RESPONSE = -2;
 const int STATUS_CHECKSUM_MISMATCH = -3;
